merges test_lpers into test_trade
