+++ tests/test_trade.py
@@ -35,6 +35,23 @@ class BaseTradeTest(unittest.TestCase):
             override_dict.update(additional_overrides)
         simulator.run_simulation(override_dict)
 
+    def run_base_lp_test(self, user_policies, config_file, additional_overrides=None):
+        """Assigns member variables that are useful for many tests"""
+        simulator = YieldSimulator(config_file)
+        simulator.set_random_variables()
+        override_dict = {
+            "pricing_model_name": "Hyperdrive",
+            "target_liquidity": 10e6,
+            "fee_percent": 0.1,
+            "init_pool_apy": 0.05,
+            "vault_apy": 0.05,
+            "num_blocks_per_day": 1,  # 1 block a day, keep it fast for testing
+            "user_policies": user_policies,  # list of user policies by name
+        }
+        if additional_overrides:
+            override_dict.update(additional_overrides)
+        simulator.run_simulation(override_dict)
+
 
 class SingleTradeTests(BaseTradeTest):
     """Tests for the SingeLong policy"""
@@ -46,3 +63,7 @@ class SingleTradeTests(BaseTradeTest):
     def test_single_short(self):
         """Tests the BaseUser class"""
         self.run_base_trade_test("single_short")
+
+    def test_base_lps(self):
+        """Tests base LP setups"""
+        self.run_base_lp_test(user_policies=["single_LP"], config_file="config/example_config.toml")
